+++ NEWS.md
@@ -1,3 +1,10 @@
+# ciphertext 0.1.1
+
+* Added bifid delastelle cipher
+* Added trifid delastelle cipher
+* decrypt=FALSE option converted in encrypt=TRUE for code robustness reasons
+
+
 # ciphertext 0.1.0
 
 * Initial CRAN submission.